requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
openpyxl==3.1.2
pandas==2.1.4
gunicorn==21.2.0
